feat: implement CRUD operations for Patient entity in PatientController and update Patient model

+++ src/main/java/dev/rm/service/PatientServiceImpl.java
@@ -1,6 +1,5 @@
 package dev.rm.service;
 
-import java.time.LocalDateTime;
 import java.util.List;
 
 import org.springframework.stereotype.Service;
@@ -8,16 +7,14 @@ import org.springframework.transaction.annotation.Transactional;
 
 import dev.rm.model.Patient;
 import dev.rm.repository.PatientRepository;
+import lombok.RequiredArgsConstructor;
 
 @Service
 @Transactional
+@RequiredArgsConstructor
 public class PatientServiceImpl implements PatientService {
 
     private final PatientRepository patientRepository;
-
-    public PatientServiceImpl(PatientRepository patientRepository) {
-        this.patientRepository = patientRepository;
-    }
 
     @Override
     public List<Patient> getAllPatients() {
@@ -37,19 +34,19 @@ public class PatientServiceImpl implements PatientService {
 
     @Override
     public Patient updatePatient(Long id, Patient patient) {
-        Patient existingPatient = patientRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("Patient not found with id: " + id));
+        Patient existingPatient = getPatientById(id);
         existingPatient.setName(patient.getName());
         existingPatient.setBirthDate(patient.getBirthDate());
         existingPatient.setDiagnosis(patient.getDiagnosis());
         existingPatient.setStatus(patient.getStatus());
-        existingPatient.setUpdatedAt(LocalDateTime.now());
         return patientRepository.save(existingPatient);
     }
 
     @Override
     public void deletePatient(Long id) {
+        if (!patientRepository.existsById(id)) {
+            throw new RuntimeException("Patient not found with id: " + id);
+        }
         patientRepository.deleteById(id);
     }
-
 }
